SwiftFormat: adjust `expandPath` to be more portable (#1468)
Reuse the implementation of `NSString`'s `isAbsolutePath` to determine
if the path is absolute rather than re-implementing that logic inline.
This allows use of `SwiftFormat` with the vknabel/SwiftFormat extension
with Visual Studio Code on Windows.

We would previously attempt to check if the configuration path was
absolute by checking for a leading `/` which does not represent an
absolute path on Windows but rather a relative path (it is drive
relative).  We perform a bit more bridging here but get the portable
behaviour without resorting to manually invoking `IsPathRelativeW` on
the path representation.  This also allows us to pick up the
improvements to the path handling in Foundation for long paths without
having to reconstruct special handling.