GHA: add a workflow for performing a Windows release (#1471)
Add some build rules to package up a Windows installer for release
builds.  Perform a separate build that is a release configuration and
does not include debug information to minimise size and optimize load
times.

Package this up into a MSI installer.  Ideally, we would perform code
signing for the MSI and the binary, but that can be done as a follow up
as it requires a Code Signing certificate to be available.

The default install location is:
  `%SystemDrive%\Program Files\nicklockwood\SwiftFormat`
following Windows conventions of
`[ProgramFiles][Manufacturer][ProductName]`.

Prefer to use `swoftprops/action-gh-release` over `actions/release` as
the latter is marked as deprecated.