Remove runtime stdlib dependency from pre-built binary targeting Linux
Currently, the pre-built binary targeting Linux links Swift stdlib and
core libraries dynamically. Unlike darwin platforms, Swift does not
guarantee ABI stability on Linux, so the pre-built binary must be run
with exactly the same version of Swift stdlib and core libraries that
it was built with on GitHub Actions. And also the shared libraries can
be installed in various locations, so users usually need to set
`LD_LIBRARY_PATH`.

However, asking users to do the above works only for power users.
Reducing the runtime dependencies makes it usable for more users.